+++ messages/ru/main.php
@@ -967,4 +967,5 @@ return
 
     "The maximum number of records per page"                          => "Допустимое количество записей на странице",
     "The minimum number of records per page"                          => "Минимально допустимое количество записей на странице",
+    "Expand/Collapse"                          => "Развернуть/Свернуть",
 ];